(#35) Fix errors in IE 8 and older
Without this patch, IE 8 raises the error "Object doesn't support this property or method" and fails to parse.  One error was caused by match() being called on a CSSStyleDeclaration object instead of a String.  Another error was caused by window.Node being undefined in IE 8 and older.  This patch checks for match() to be a function before calling it.  This patch also defines window.Node if it needs to be defined.