Rename anchor plugin
The `europa-plugin-anchor` has been renamed to `europa-plugin-link` in order to better reflect the proposed naming standards for plugins. The documentation has also been updated to better reflect this.

At the same title as the "anchor" footnote references were renamed to "link"; footnote references are no longer zero-based and now start at `1` for both links and images.