Simplify recording references
The `europa-plugin-image` and `europa-plugin-link` packages shared a lot of logic relating to the recording and rendering of Markdown references at the end of the output.

A `Conversion#addReference` method has been added to drastically simplify this and the `Conversion#end` method now ensures any record references are rendered. The new method must only be called when the `inline` option is disabled.