Replace JSDom with Cheerio
This commit is huge, sorry. It definitely didn't need to be but it turned out that way and, given the huge refactoring going on recently, I just went with it.

The core of this commit is getting rid of the `jsdom` dependency for Node.js and replacing it with `cheerio`. Due to the fact that this removed any notion of a DOM (even virtual) from Node.js, causing everything to break. To solve this, the concept of a DOM wrapper has been introduced and `europa-core` implementations are now required to provide one. It is used by `europa-core` and all plugins to interact with the DOM indirectly. Now that there is no virtual DOM support, the `release` method has been removed from the API along with the `document` and `window` properties, greatly simplifying the core API.

This change also resulted in a relatively empty `WindowService` which really served no real purpose as a standalone service but didn't make sense to merge with the new `CharsetService` either. The decision has been made to consolidate both services into a new concept; an `Environment`. This is now a wrapper for all environment-specific information that a `europa-core` implementation must provide, including the new DOM wrapper. Two new packages have been created as a result; `europa-environment-browser` and `europa-environment-node`. The hope here is that these could be reused for other implementations and/or for testing plugins in isolation of a specific `europa-core` implementation and any other plugins.

A hook has now been added to allow plugins to control how the value of text DOM nodes are converted, abstracting more plugin-specific logic from the core into the plugins.